about to begin porting tests
Moved current tests to new file so I can systematically go through the tests and ensure non are lost
